# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "DBMS",
        "version": "0.1.0"
    },
    "paths": {
        "/databases": {
            "get": {
                "summary": "Get Databases",
                "operationId": "get_databases_databases_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "string"
                                    },
                                    "type": "array",
                                    "title": "Response Get Databases Databases Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Database",
                "operationId": "create_database_databases_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewDatabase"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Database"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Database already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                },
                                "example": {
                                    "detail": "Database 'database_name' already exists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/databases/{database_name}": {
            "get": {
                "summary": "Get Database",
                "operationId": "get_database_databases__database_name__get",
                "parameters": [
                    {
                        "name": "database_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Database Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Database"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Database was not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "detail": "Cannot find database 'database_name'"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Database",
                "operationId": "delete_database_databases__database_name__delete",
                "parameters": [
                    {
                        "name": "database_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Database Name"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "404": {
                        "description": "Database was not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "detail": "Cannot find database 'database_name'"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/databases/{database_name}/tables/{table_id}": {
            "get": {
                "summary": "Get Table",
                "operationId": "get_table_databases__database_name__tables__table_id__get",
                "parameters": [
                    {
                        "name": "database_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Database Name"
                        }
                    },
                    {
                        "name": "table_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Table Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Table"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Database or table was not found",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "database_not_found": {
                                        "summary": "database not found",
                                        "value": {
                                            "detail": "Cannot find database 'database_name'"
                                        }
                                    },
                                    "table_not_found": {
                                        "summary": "table not found",
                                        "value": {
                                            "detail": "Database 'database_name' doesn't contain table #table_id"
                                        }
                                    }
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Table",
                "operationId": "delete_table_databases__database_name__tables__table_id__delete",
                "parameters": [
                    {
                        "name": "database_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Database Name"
                        }
                    },
                    {
                        "name": "table_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Table Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "404": {
                        "description": "Database or table was not found",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "database_not_found": {
                                        "summary": "database not found",
                                        "value": {
                                            "detail": "Cannot find database 'database_name'"
                                        }
                                    },
                                    "table_not_found": {
                                        "summary": "table not found",
                                        "value": {
                                            "detail": "Database 'database_name' doesn't contain table #table_id"
                                        }
                                    }
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/databases/{database_name}/tables/": {
            "post": {
                "summary": "Create Table",
                "operationId": "create_table_databases__database_name__tables__post",
                "parameters": [
                    {
                        "name": "database_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Database Name"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewTable"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Table"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Database was not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "detail": "Cannot find database 'database_name'"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/databases/{database_name}/tables/{table_id}/rows/{row_id}": {
            "get": {
                "summary": "Get Row",
                "operationId": "get_row_databases__database_name__tables__table_id__rows__row_id__get",
                "parameters": [
                    {
                        "name": "database_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Database Name"
                        }
                    },
                    {
                        "name": "table_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Table Id"
                        }
                    },
                    {
                        "name": "row_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Row Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Row"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Database, table or row was not found",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "database_not_found": {
                                        "summary": "database not found",
                                        "value": {
                                            "detail": "Cannot find database 'database_name'"
                                        }
                                    },
                                    "table_not_found": {
                                        "summary": "table not found",
                                        "value": {
                                            "detail": "Database 'database_name' doesn't contain table #table_id"
                                        }
                                    },
                                    "row_not_found": {
                                        "summary": "row not found",
                                        "value": {
                                            "detail": "Table #table_id in database 'database_name' doesn't contain row #row_id"
                                        }
                                    }
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Row",
                "operationId": "delete_row_databases__database_name__tables__table_id__rows__row_id__delete",
                "parameters": [
                    {
                        "name": "database_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Database Name"
                        }
                    },
                    {
                        "name": "table_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Table Id"
                        }
                    },
                    {
                        "name": "row_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Row Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "404": {
                        "description": "Database, table or row was not found",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "database_not_found": {
                                        "summary": "database not found",
                                        "value": {
                                            "detail": "Cannot find database 'database_name'"
                                        }
                                    },
                                    "table_not_found": {
                                        "summary": "table not found",
                                        "value": {
                                            "detail": "Database 'database_name' doesn't contain table #table_id"
                                        }
                                    },
                                    "row_not_found": {
                                        "summary": "row not found",
                                        "value": {
                                            "detail": "Table #table_id in database 'database_name' doesn't contain row #row_id"
                                        }
                                    }
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Database": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "tables": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/Table"
                        },
                        "type": "object",
                        "title": "Tables",
                        "default": {}
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "Database"
            },
            "HTTPError": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "HTTPError"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "NewDatabase": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "NewDatabase"
            },
...
            "NewTable": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "columns": {
                        "items": {
                            "$ref": "#/components/schemas/Column"
                        },
                        "type": "array",
                        "title": "Columns"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "columns"
                ],
                "title": "NewTable"
            },
...
            "Table": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "columns": {
                        "items": {
                            "$ref": "#/components/schemas/Column"
                        },
                        "type": "array",
                        "title": "Columns"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "columns"
                ],
                "title": "Table"
            },
...
        }
    }
}